Stories adjustments - Input colors - NumberInput - added variant and styles for prefix - Form - moved storybook code to separate variable to display the whole behaviour
