+++ src/components/Form/Form.stories.tsx
@@ -1,17 +1,12 @@
 import React from 'react';
-import { ComponentStory, ComponentMeta } from '@storybook/react';
+import { ComponentMeta } from '@storybook/react';
 import Form from './index';
-import Input from '../Input';
-import InputPassword from '../InputPassword';
-import * as yup from 'yup';
-import Button from '../Button';
-import Multiselect from '../Multiselect';
-import Select from '../Select';
-import Combobox from '../Combobox';
-import { defaultTheme } from '../Theme';
 import pkg from './package.json';
-import FormDisabledSubmitButton from '../../storybook/preview/FormDisabledSubmitButton';
-import ExternalFormControl from '../../storybook/preview/ExternalFormControl';
+import FormDisabledSubmitButton, { FormDisabledSubmitButtonCode } from '../../storybook/preview/Form/DisabledSubmitButton';
+import FormExternalFormControl, { FormExternalControlCode } from '../../storybook/preview/Form/ExternalControl';
+import FormValidation, { FormValidationCode } from '../../storybook/preview/Form/Validation';
+import FormInitialValues, { FormInitialValuesCode } from '../../storybook/preview/Form/InitialValues';
+import FormSetErrorOnSubmit, { FormSetErrorOnSubmitCode } from '../../storybook/preview/Form/SetErrorOnSubmit';
 
 export default {
   title: 'Components/Form',
@@ -21,91 +16,47 @@ export default {
   },
 } as ComponentMeta<typeof Form>;
 
-const Template: ComponentStory<typeof Form> = (args) => {
-  const { onValidate, onValidateAsync, ...props } = args;
-
-  return <Form {...props} />;
-};
-
-const FormContent = ({ title, text, showSelects = true }: {title: string, showSelects?: boolean, text?: string}) => (
-  <>
-    <h2 style={{ fontFamily: 'Inter', color: defaultTheme.primary.main }}>{title}</h2>
-    {text && <h4 style={{ fontFamily: 'Inter', color: defaultTheme.text.hint, fontWeight: 500 }}>{text}</h4>}
-    <Input label={'Email'} name="email" placeholder="some@mail.com"/>
-    <h1> </h1>
-    <InputPassword label={'Password'} name="password" placeholder="Password is required"/>
-    <h1> </h1>
-    {showSelects && (
-      <>
-        <Select
-          name="select"
-          options={['One', 'Two', 'Three']}
-          label="Select"
-          placeholder="Select a value"
-        />
-        <h1> </h1>
-        <Multiselect
-          name="multiselect"
-          options={['One', 'Two', 'Three']}
-          label="Multiselect"
-          placeholder="Select at least one value"
-        />
-        <h1> </h1>
-        <Combobox
-          name="combobox"
-          options={[{ label: 'One', value: 1 }, { label: 'Two', value: 2 }, { label: 'Three', value: 3 }]}
-          label="Combobox"
-          placeholder="Select a value"
-        />
-        <h1> </h1>
-      </>
-    )}
-    <div style={{ display: 'flex', gap: 20 }}>
-      <Button label="Submit" type="submit"/>
-      <Button label="Reset" type="reset" variant="outlined"/>
-    </div>
-  </>
-);
-
-export const Validation = Template.bind({});
-Validation.args = {
-  onSubmit: () => {
-    alert('Submitted')
+export const InitialValues = () => <FormInitialValues />;
+InitialValues.parameters = {
+  docs: {
+    source: {
+      code: FormInitialValuesCode
+    },
   },
-  children: <FormContent title="Form with validation" />,
-  validationSchema: yup.object({
-    email: yup.string().email().required(),
-    password: yup.string().required(),
-    select: yup.string().required(),
-    combobox: yup.string().required(),
-    multiselect: yup.array().required().min(1),
-  }),
 };
 
-export const InitialValues = Template.bind({});
-InitialValues.args = {
-  onSubmit: () => {
-    alert('Submitted')
+export const Validation = () => <FormValidation />;
+Validation.parameters = {
+  docs: {
+    source: {
+      code: FormValidationCode
+    },
   },
-  children: <FormContent title="Form with initial values" />,
-  initialValues: {
-    email: 'somemail@mail.com',
-    password: 'TGBwfe23',
-    select: 'Two',
-    multiselect: ['Two', 'Three'],
-    combobox: 1,
-  }
 };
 
-export const ErrorOnSubmit = Template.bind({});
-ErrorOnSubmit.args = {
-  onSubmit: (formData: any, { setErrors }) => {
-    setErrors({ email: 'This email is already taken' });
+export const ErrorOnSubmit = () => <FormSetErrorOnSubmit />;
+ErrorOnSubmit.parameters = {
+  docs: {
+    source: {
+      code: FormSetErrorOnSubmitCode,
+    },
   },
-  children: <FormContent title="Form error after submit" showSelects={false}
-    text="setError Form action is used instead of validation"/>,
 };
 
-export const DisabledSubmitButton = FormDisabledSubmitButton;
+export const DisableSubmitButton = () => <FormDisabledSubmitButton />;
+DisableSubmitButton.parameters = {
+  docs: {
+    source: {
+      code: FormDisabledSubmitButtonCode,
+    },
+  },
+};
 
-export const ExternalControl = ExternalFormControl;
+export const ExternalControl = () => <FormExternalFormControl />;
+ExternalControl.parameters = {
+  docs: {
+    source: {
+      code: FormExternalControlCode,
+    },
+  },
+};
