+++ src/components/NumberInput/style.css
@@ -1,0 +1,3 @@
+.prefix {
+  margin-right: 5px;
+}
